+++ include/maxtree_node.hpp
@@ -9,12 +9,12 @@ enum maxtee_node_field {
 
 class maxtree_node{
     public:
-        int parent;
-        int label;
-        unsigned int idx;
+        long long int parent;
+        long long int label;
+        unsigned long long int idx;
         double gval;
 
-        maxtree_node(double g, unsigned int i);
+        maxtree_node(double g, unsigned long long int i);
     
 
         bool operator>(const maxtree_node &r);
